+++ tsup.config.ts
@@ -2,11 +2,11 @@ import { defineConfig } from "tsup";
 
 export default defineConfig((options) => {
   return {
-    entry: ["src/index.ts", "src/components/RFloatingPopup/index.ts"],
+    entry: ["src/index.ts"],
     dts: true,
     format: ["esm"],
     esbuildOptions(options) {
-      options.external = ["react", "react-dom", "maplibre-gl", "@floating-ui/dom", "clsx"];
+      options.external = ["react", "react-dom", "maplibre-gl", "clsx"];
     },
     outExtension: () => {
       return {